requirements.txt
django>=3.2.0
firebase-admin>=6.2.0
djangorestframework>=3.14.0
